+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented here. The format is based
 - Added "OpenSim" muscle coloring option, which uses `GeometryPath::getColor` (#193)
 - Added ImPlot demo screen to the experimental screens section (#194)
 - Added ability to open multiple muscle plots in the editor (related: #191)
+- Added ability to change a coordinate value via a muscle plot by clicking on the muscle plot (#195)
 
 
 ## [0.1.5] - 2022/04/11